backend/requirements.txt
# --- Web ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1

# --- Auth / segurança ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- DB ---
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# --- Storage ---
boto3==1.35.79

# --- LLM / Embeddings ---
anthropic==0.40.0
openai==1.57.0

# --- Vector DB ---
chromadb==0.5.23

# --- OCR / PDF ---
pdfplumber==0.11.4
PyMuPDF==1.24.14
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0

# --- Grafo ---
networkx==3.4.2

# --- Util ---
httpx==0.28.1
python-slugify==8.0.4
tenacity==9.0.0
